Move vocab thing to vocab.py
